🔧 Fix critical import and NumPy type issues
✅ FIXED CRITICAL IMPORT ERROR:
- Restored missing 'pickle' import needed for cache operations
- Error was: "Failed to save cache: name 'pickle' is not defined"
- Cache operations were failing silently, potentially causing data issues

✅ FIXED NUMPY TYPE SERIALIZATION:
- Fixed np.mean() and np.median() returning NumPy types instead of Python floats
- NumPy types can cause "Undefined __round__ method" errors in JavaScript/JSON contexts
- Added explicit float() conversion to ensure Python native types

✅ ROOT CAUSE ANALYSIS:
- During import cleanup, accidentally removed required 'pickle' import
- NumPy types from sophisticated analysis were not being converted to JSON-safe types
- Both issues could cause serialization/round() method errors

✅ SOLUTION:
- Restored pickle import for cache functionality
- Convert np.mean()/np.median() results to Python floats explicitly

+++ sophisticated_timeframe.py
@@ -343,8 +343,8 @@ class SophisticatedTimeframePredictor:
                         })
             
             if recoveries:
-                patterns['avg_recovery_days'] = round(np.mean(recoveries), 1)
-                patterns['median_recovery_days'] = round(np.median(recoveries), 1)
+                patterns['avg_recovery_days'] = round(float(np.mean(recoveries)), 1)
+                patterns['median_recovery_days'] = round(float(np.median(recoveries)), 1)
                 patterns['fastest_recovery'] = min(recoveries)
                 patterns['historical_success_rate'] = round((len(recoveries) / max(1, len([d for d in drawdowns if d < -3]))) * 100, 1)
